Save resized bitmap onDisk
This for some reason is a lot faster than storing it in memory and passing it
on to wallpaper manager via setBitmap. My guess is due to different formats, ie
us providing images via jpeg and wallpaper manager insisting on converting it to png,
is what causes the long decode/compress times.